+++ chutes/constants.py
@@ -1,1 +1,9 @@
 CHUTES_DIR = ".chutes"
+HOTKEY_HEADER = "X-Parachutes-Hotkey"
+SIGNATURE_HEADER = "X-Parachutes-Signature"
+NONCE_HEADER = "X-Parachutes-Nonce"
+AUTHORIZATION_HEADER = "Authorization"
+USER_ID_HEADER = "X-Parachutes-UserID"
+CHUTEID_HEADER = "X-Parachutes-ChuteID"
+FUNCTION_HEADER = "X-Parachutes-Function"
+